Organize Files into Folders
Put pages in page folder, task data model in model folder, and database/utility functions into database folder
